back_project/requirementsDMP.txt
aiohttp==3.8.0
aiosignal==1.2.0
alembic==1.7.4
aniso8601==9.0.1
argcomplete==1.12.3
async-timeout==4.0.0a3
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
biopython==1.79
boto3==1.18.11
botocore==1.21.11
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.3
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.3.0
debugpy==1.4.1
decorator==5.0.9
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flex==6.14.1
frozenlist==1.2.0
greenlet==1.1.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipinfo==4.2.1
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.0.1
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
keyring==23.2.1
keyrings.alt==4.1.0
lxml==4.6.4
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
multidict==5.2.0
mysqlclient==2.0.3
nest-asyncio==1.5.1
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
paramiko==2.8.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
premailer==3.10.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.7.6
requests==2.26.0
rfc3987==1.3.8
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.7.1
SecretStorage==3.3.1
semantic-version==2.8.5
shortuuid==1.0.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.5
strict-rfc3339==0.7
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.0.4
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
traitlets==5.1.0
twilio==6.61.0
typing-extensions==3.10.0.2
urllib3==1.26.6
validate-email==1.3
wcwidth==0.2.5
websocket-client==0.59.0
Werkzeug==1.0.1
wget==3.2
xlrd==2.0.1
XlsxWriter==1.4.5
yagmail==0.14.260
yarl==1.7.0
zipp==3.5.0
